refactor: SingleImageProvider move pacakge
to io.github.vinceglb.filekit.dialogs.util
